@@ -114,7 +114,7 @@ class capture(emulator):
         filename = self._file_template.format(self._count)
         image = self.preprocess(image)
         surface = self.to_surface(image, alpha=self._contrast)
-        logger.debug("Writing: {0}".format(filename))
+        logger.debug(f"Writing: {filename}")
         self._pygame.image.save(surface, filename)
 
 
